mapek/requirements.txt
psycopg2-binary>=2.9.9
pydantic>=2.6.4
structlog>=24.1.0
prometheus_client>=0.19.0
redis>=5.0.4
tenacity>=8.2.3
aiohttp>=3.9.5
pytest>=8.2.2
numpy>=1.26.4
aiohttp
tenacity
pybreaker>=1.0.0